[CP-3170] useAbortFlashingOnDeviceDetached -> useMscDeviceDetachedEffect (with debounce logic)
