+++ libs/core/core/components/apps/base-app/base-app.component.tsx
@@ -16,7 +16,7 @@ import { useDeviceLockedEffect } from "Core/core/hooks/use-device-locked-effect"
 import { useDeviceDetachedEffect } from "Core/core/hooks/use-device-detached-effect"
 import { useDeviceConnectFailedEffect } from "Core/core/hooks/use-device-connect-failed-effect"
 import { useDiscoveryRedirectEffect } from "Core/core/hooks/use-discovery-redirect-effect"
-import { useAbortFlashingOnDeviceDetached } from "Core/core/hooks/use-abort-flashing-on-device-detached"
+import { useMscDeviceDetachedEffect } from "Core/core/hooks/use-msc-device-detached-effect"
 import { useRouterListener } from "Core/core/hooks"
 import {
   OutboxWrapper,
@@ -48,7 +48,7 @@ const BaseApp: FunctionComponent = () => {
   useHelp()
 
   // MSC
-  useAbortFlashingOnDeviceDetached()
+  useMscDeviceDetachedEffect()
 
   return (
     <>
